requirements.txt
yfinance
pandas
google-generativeai
requests-cache
